Create README.md
This is an Atm interface that uses java classes, objects, it results in the successful transfer, deposition, withdrawal of money.

A number Guessing game that uses java, in which there are three choices play, score, exit. enter any choice to proceed. you have the limit of the number and it also tells whether the entered number is higher or lower than the correct number.